% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createSearchFromTags}
\alias{createSearchFromTags}
\title{Convert a list of hashtags to a Twitter search string}
\usage{
createSearchFromTags(ht, op = "OR")
}
\arguments{
\item{ht}{the list of tags created using (e.g.) hashTags <- c("tag1", "tag2")}

\item{op}{how to combine hashtag strings, default = OR}
}
\description{
It's usually easier to keep the hashtags you want to search for in an R list. This function
creates a proper API search using the operator you specify.
}
\details{
Do not include # in the list elements as this will be added automatically
}
\seealso{
Other tweets: \code{\link{getHashtags}},
  \code{\link{loadTweets}}, \code{\link{processTweets}},
  \code{\link{saveTweets}}
}
\author{
Ben Anderson, \email{banderson@soton.ac.uk}
}
\concept{tweets}
