% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processTweets.R
\name{getHashtags}
\alias{getHashtags}
\title{Extract a list of unique hashtags}
\usage{
getHashtags(dt, keepVars = c("user_id", "status_id", "screen_name",
  "created_at", "ba_tweetType", "hashtags"))
}
\arguments{
\item{dt}{data.table of tweets}

\item{keep}{a list of the variables to keep. default = c("user_id", "status_id", "screen_name", "created_at", "ba_tweetType", "hashtags")}
}
\description{
\code{getHashTags} takes a data.table of processed tweets and creates a new table of unique hashtags by created_at
}
\details{
Returns a data.table
}
\seealso{
Other tweets: \code{\link{createSearchFromTags}},
  \code{\link{loadTweets}}, \code{\link{processTweets}},
  \code{\link{saveTweets}}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{tweets}
