% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweets.R
\name{getTweets}
\alias{getTweets}
\title{Gets tweets that match a given string}
\usage{
getTweets(string, n = 18000)
}
\arguments{
\item{string}{the string to look for}

\item{n}{number of tweets, default = 18000 (see https://github.com/mkearney/rtweet)}

\item{ofile}{file to save to}
}
\description{
\code{getTweets} takes a string and matches it to recent tweets. It includes retweets.
}
\details{
Puts the results into a data.table and returns the data.table for further fun.
}
\seealso{
Other tweets: 
\code{\link{createSearchFromTags}()},
\code{\link{getHashtags}()},
\code{\link{loadTweets}()},
\code{\link{processTweets}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{tweets}
