% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{loadTweets}
\alias{loadTweets}
\title{Loads tweets that match a given search string from a pre-saved file}
\usage{
loadTweets(path, string)
}
\arguments{
\item{path}{place to look for saved file - gets passed to readr::read_csv so can be anything read_csv can readr :-)}

\item{string}{the string to look for}
}
\description{
\code{loadTweets} takes a string and loads all files in a given path which have this name using readr
}
\details{
Assumes (but checks) files were previously created using saveTweets

Puts the results into a tbl and returns them.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
