% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processTweets.R
\name{processTweets}
\alias{processTweets}
\title{Process a raw tweets file}
\usage{
processTweets(dt)
}
\arguments{
\item{dt}{data.table of tweets to clean}
}
\description{
\code{processTweets} takes a data.table of tweets returned from the API and read using readTweets
}
\details{
Removes exact duplicates
   
   Creates a date and hms from the created_at column
   
   Returns a cleaned data.table
}
\seealso{
Other tweets: 
\code{\link{createSearchFromTags}()},
\code{\link{getHashtags}()},
\code{\link{getTweets}()},
\code{\link{loadTweets}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{tweets}
