% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{saveTweets}
\alias{saveTweets}
\title{Gets tweets that match a given string and stores them}
\usage{
saveTweets(string, ofile, n = 18000)
}
\arguments{
\item{string}{the string to look for}

\item{ofile}{file to save to}

\item{n}{number of tweets, default = 18000 (see https://github.com/mkearney/rtweet)}
}
\description{
\code{saveTweets} takes a string and matches it to recent tweets
}
\details{
Puts the results into a data.table and saves it to ofile using data.table::fwrite (very fast). Gzips the file.

Returns the data.table for further fun.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
