% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putData.R
\name{sendDM}
\alias{sendDM}
\title{Sends a DM from the app}
\usage{
sendDM(text, user)
}
\arguments{
\item{string}{the string to look for}

\item{ofile}{file to save to}

\item{n}{number of tweets, default = 18000 (see https://github.com/mkearney/rtweet)}
}
\description{
\code{sendDM} takes a string and matches it to recent tweets
}
\details{
Puts the results into a tbl and saves it to ofile. Returns the tbl for further fun.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
