% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{loadTweets}
\alias{loadTweets}
\title{Loads tweets that match a given search string from a pre-saved file}
\usage{
loadTweets(path, pattern, recursive = TRUE)
}
\arguments{
\item{path}{place to look for saved file - gets passed to list.files()}

\item{pattern}{the pattern to match - passed to list.files()}

\item{recursive}{use recursively (default = TRUE)}
}
\description{
\code{loadTweets} takes a string and loads all files in a given path which have this name using readr
}
\details{
If no matches, returns nothing but gives feedback.

If there are matches, rbinds them all into a tbl, gives feedback and returns the tbl as a data.table
  for further fun.

Note that list.files() is recursive by default. This can lead to large search times if you start it at
  the top of a file structure. Use wisely.
}
\seealso{
Other tweets: \code{\link{createSearchFromTags}},
  \code{\link{saveTweets}}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{tweets}
